+++ NEWS.md
@@ -6,6 +6,7 @@
 * All functions that take table name as argument now work with
   `dbplyr::in_schema()` and `DBI::Id()` besides the usual character table name.
   (#228)
+* Create a new `dbRenameTable()` function to rename table name.
 
 # RPresto 1.4.1